refactor (test/services): Ts'ifying test/services
Translates AtScript files in test/services to TypeScript
