+++ modules/angular2/test/services/rectangle_mock.ts
