+++ tools/broccoli/trees/node_tree.ts
@@ -23,7 +23,7 @@ module.exports = function makeNodeTree(destinationPath) {
       // the following code and tests are not compatible with CJS/node environment
       'angular2/test/core/zone/**',
       'angular2/test/test_lib/fake_async_spec.js',
-      'angular2/test/services/xhr_impl_spec.js'
+      'angular2/test/services/xhr_impl_spec.ts'
     ]
   });
